+++ cmd/template/cluster/provider/templates/capa/types.go
@@ -19,6 +19,7 @@ type Metadata struct {
 	Name            string            `json:"name,omitempty"`
 	Organization    string            `json:"organization,omitempty"`
 	PreventDeletion bool              `json:"preventDeletion,omitempty"`
+	ServicePriority string            `json:"servicePriority,omitempty"`
 }
 
 type Release struct {