+++ cmd/template/cluster/provider/templates/capv/types.go
@@ -18,6 +18,7 @@ type Metadata struct {
 	Name            string            `json:"name,omitempty"`
 	Organization    string            `json:"organization,omitempty"`
 	PreventDeletion bool              `json:"preventDeletion,omitempty"`
+	ServicePriority string            `json:"servicePriority,omitempty"`
 }
 
 type DefaultAppsConfig struct {