+++ cmd/template/cluster/provider/capv.go
@@ -159,10 +159,7 @@ func BuildCapvClusterConfig(config common.ClusterConfig) capv.ClusterConfig {
 				},
 			},
 			ControlPlane: &capv.ControlPlane{
-				Replicas: config.VSphere.ControlPlane.Replicas,
-				Image: &capv.Image{
-					Repository: "gsoci.azurecr.io/giantswarm",
-				},
+				Replicas:        config.VSphere.ControlPlane.Replicas,
 				MachineTemplate: getMachineTemplate(&config.VSphere.ControlPlane.VSphereMachineTemplate, &config),
 			},
 			Metadata: &capv.Metadata{